+++ packages/examples/src/demo/init_data/logic/with_func_in_lhs.js
@@ -1,0 +1,39 @@
+export default
+{
+  "and": [
+    {
+      "!": {
+        "in": [
+          "Den",
+          {
+            "toLowerCase": [
+              {
+                "var": "user.firstName"
+              }
+            ]
+          }
+        ]
+      }
+    },
+    {
+      "!=": [
+        {
+          "toLowerCase": [
+            {
+              "toUpperCase": [
+                {
+                  "var": "user.login"
+                }
+              ]
+            }
+          ]
+        },
+        {
+          "toLowerCase": [
+            "user"
+          ]
+        }
+      ]
+    }
+  ]
+};